Configure Git LFS for large files (#12)
There will be a video overview in the documentation and while it's maybe not the best idea to store video files in git, we'll use git lfs to decrease some of the downsides. The key reason for doing this is that it allows us to make changes to the video and not have to worry about getting a new YouTube URL and then possibly having people who referred to the old URL, etc.

(Rules bypassed because there is no spell check or format check for this PR. Will address this rules scenario in a future change.)